src/snagrecover/utils.py: Create a generic Header class
Move the Header class from the zynqmp_fw.py into snagrecover/utils.py
to allow reusing this class elsewhere.
