+++ src/snagrecover/firmware/zynqmp_fw.py
@@ -5,28 +5,9 @@ Handle ZynqMP boot images, as generated by the Xilinx bootgen tool.
 import logging
 logger = logging.getLogger("snagrecover")
 from snagrecover.protocols import dfu
+from snagrecover.utils import BinFileHeader as Header
 from dataclasses import dataclass, astuple
 import struct
-
-class Header():
-	"""
-	Generic class representing a packed C struct embedded in a bytearray.
-	"""
-
-	fmt = ""
-	size = 0
-
-	@classmethod
-	def read(cls, data, offset=0):
-		obj = cls(*struct.unpack(cls.fmt, data[offset:offset + cls.size]))
-		obj.offset = offset
-
-		return obj
-
-	@classmethod
-	def write(cls, self, data):
-		offset = self.offset
-		data[offset:offset + cls.size] = struct.pack(cls.fmt, *astuple(self))
 
 @dataclass
 class ZynqMPImageTable(Header):
@@ -40,7 +21,7 @@ class ZynqMPImageTable(Header):
 	csum: int
 
 	fmt = "<LLLLLL36sL"
-	size = 0x40
+	class_size = 0x40
 
 	def update_checksum(self):
 		data = struct.pack(ZynqMPImageTable.fmt, *astuple(self))
@@ -55,7 +36,7 @@ class ZynqMPImageHeader(Header):
 	# Other fields are ignored
 
 	fmt = "<LLLL"
-	size = 16
+	class_size = 16
 
 @dataclass
 class ZynqMPPartHeader(Header):
@@ -77,7 +58,7 @@ class ZynqMPPartHeader(Header):
 	csum: int
 
 	fmt = "<LLLLLLLLLLLLLLLL"
-	size = 0x40
+	class_size = 0x40
 
 	def update_checksum(self):
 		data = struct.pack(ZynqMPPartHeader.fmt, *astuple(self))
